security

+++ app/Http/Controllers/Letter/FolderController.php
@@ -14,7 +14,7 @@ class FolderController extends Controller
      */
     public function index()
     {
-        $folders = Folder::all();
+        $folders = Folder::currentoffice()->get();
         return inertia('letter/folder/Index', [
             'folders' => $folders,
         ]);
@@ -75,13 +75,18 @@ class FolderController extends Controller
      */
     public function edit(string $id)
     {
+        $folder = Folder::findOrFail($id);
+        abort_if(
+            auth()->user()->role == 'user' && auth()->user()->office_id !== $folder->office_id,
+            403,
+            "You are not auhorized to edit the folder."
+        );
         $offices = Office::all()->map(function ($office) {
             $office->office_name = json_decode($office->office_name, true);
             $office->upazila = json_decode($office->upazila, true);
             $office->district = json_decode($office->district, true);
             return $office;
         });
-        $folder = Folder::findOrFail($id);
         return inertia('letter/folder/Edit', [
             'folder' => $folder,
             'offices' => $offices,
@@ -94,6 +99,11 @@ class FolderController extends Controller
     public function update(Request $request, string $id)
     {
         $folder = Folder::findOrFail($id);
+        abort_if(
+            auth()->user()->role == 'user' && auth()->user()->office_id !== $folder->office_id,
+            403,
+            "You are not auhorized to edit the folder."
+        );
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'description' => ['nullable', 'string'],
@@ -121,6 +131,11 @@ class FolderController extends Controller
     public function destroy(string $id)
     {
         $folder = Folder::findOrFail($id);
+        abort_if(
+            auth()->user()->role == 'user' && auth()->user()->office_id !== $folder->office_id,
+            403,
+            "You are not auhorized to edit the folder."
+        );
         $folder->delete();
         $toast = [
             'message' => 'Folder deleted successfully.',
